renamed style files
