+++ client/src/components/AuthHeader/authHeaderStyles.ts
