+++ client/src/components/ChatSideBanner/chatSideBarStyles.ts
