+++ client/src/components/Search/searchStyles.ts
