+++ client/src/pages/SignUp/signUpStyles.ts
